requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-jwt-2fa==0.3.0
Markdown==3.2.2
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
pysftp==0.2.9
sqlparse
drf-jwt-2fa
django-url-filter
django_dbconn_retry